+++ NEWS.md
@@ -1,5 +1,7 @@
 # roxygen2 3.0.0.99
 
+* Don't try and add parameters for data objects (#165)
+
 # roxygen2 3.0.0
 
 Roxygen2 now fully supports S4 and RC (reference classes) - you should no